backend/playground/requirements.txt
reportlab==4.0.7
google-cloud-documentai==2.20.1
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.1.0
PyMuPDF==1.23.8
docx2pdf==0.1.8
python-docx==1.1.0 